Complete SearchRouter renaming
