+++ app/search-router.js
@@ -1,11 +1,11 @@
 import 'setimmediate';
 import 'regenerator-runtime';
 import asyncHandler from 'express-async-handler';
-import QueryState from '../query-state.js';
-import Sidebar from '../sidebar.js';
-import Searcher from '../searcher.js';
+import QueryState from './query-state.js';
+import Sidebar from './sidebar.js';
+import Searcher from './searcher.js';
 
-export const search = asyncHandler(async (req, res) => {
+const search = asyncHandler(async (req, res) => {
   const queryState = new QueryState(req);
   const sidebar = await Sidebar.getSidebar(queryState.query);
   queryState.results = await Searcher.search(queryState);
@@ -15,3 +15,5 @@ export const search = asyncHandler(async (req, res) => {
 
   res.render('search', { queryState, sidebar });
 });
+
+export default { search };
